requirements.txt
openpyxl>=3.1.0
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=10.0.0
filelock>=3.12
imagehash>=4.3
flask>=2.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
anthropic>=0.21.0
